implement redux api middleware pattern for api calls
