+++ frontend/app/actions/catActions.js
@@ -1,0 +1,24 @@
+import { RSAA } from 'redux-api-middleware'
+
+export const REQUEST = 'RANDOM_CAT_REQUEST'
+export const SUCCESS = 'RANDOM_CAT_SUCCESS'
+export const FAILURE = 'RANDOM_CAT_FAILURE'
+
+export function fetchRandomCat() {
+  return {
+    [RSAA]: {
+      endpoint: '/api/pet.getRandom?output=full',
+      method: 'GET',
+      types: [
+        REQUEST,
+        {
+          type: SUCCESS,
+          payload: (action, state, res) => {
+            return res.json().then(json => json['petfinder']['pet'])
+          }
+        },
+        FAILURE
+      ]
+    }
+  }
+}
