+++ package.json
@@ -43,6 +43,7 @@
     "react-hot-loader": "^3.1.3",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
+    "redux-api-middleware": "^2.2.0",
     "redux-devtools": "^3.4.1",
     "redux-thunk": "^2.2.0",
     "sass-loader": "^6.0.6",
